Refactoring

+++ controllers/IdeaController.js
@@ -2,18 +2,21 @@ const ideaDao = require('../database/idea.dao');
 
 const postIdea = async (req, res) => {
   req.body.creator = req.decoded.userEmail;
-  req.body.average_score = (req.body.impact + req.body.ease + req.body.confidence) / 3.
+  req.body.average_score = (req.body.impact + req.body.ease + req.body.confidence) / 3.0;
   req.body.created_at = Date.now();
-  const idea = await ideaDao.createIdea(req.body);
+  let idea = await ideaDao.createIdea(req.body);
   if (idea) {
-    idea.__v = undefined;
-    idea.creator = undefined;
-    idea._id = undefined;
-    res.status(200).send(idea);
+    let finalIdea = {
+      id: idea._id,
+      ...idea._doc,
+      _id: undefined,
+      __v : undefined,
+      creator :undefined,
+    }
+    res.status(200).send(finalIdea);
   } else {
-    res.status(500).send("Error creating the idea");
+    res.status(500).send('Error creating the idea');
   }
-
 };
 
 const deleteIdea = async (req, res) => {
@@ -21,8 +24,17 @@ const deleteIdea = async (req, res) => {
     res.status(404).send();
   } else {
     const { id } = req.params;
-    await ideaDao.deleteIdea(id);
-    res.status(204).send();
+    const idea = await ideaDao.getIdea(id);
+    if (idea) {
+      if (idea.creator === req.decoded.userEmail) {
+        await ideaDao.deleteIdea(id);
+        res.status(204).send();
+      } else {
+        res.status(401).send();
+      }
+    } else {
+      res.status(404).send('Idea not found');
+    }
   }
 };
 
@@ -31,24 +43,48 @@ const getIdeas = async (req, res) => {
   const pageSize = 10;
   const pageOffset = (page * pageSize) - pageSize;
   const ideas = await ideaDao.getIdeas(req.decoded.userEmail, pageOffset, pageSize);
-  ideas.forEach(idea=>{
-    idea["__v"] = undefined;
-    idea["creator"] = undefined;
-  })
-  res.status(200).send(ideas);
+  if(ideas){
+    let finalIdeas = ideas.map(function(idea) {
+      console.log(idea);
+      return  {
+        id: idea._id,
+        ...idea._doc,
+        _id: undefined,
+        __v : undefined,
+        creator :undefined,
+      }
+    });
+    res.status(200).send(finalIdeas);
+  }else{
+    res.status(404).send();
+  }
 };
 
 const updateIdea = async (req, res) => {
-  console.log("Req", req.params)
   if (!req.params.id) {
     res.status(404).send();
   } else {
-    req.body.average_score = (req.body.impact + req.body.ease + req.body.confidence) / 3.
-    const updatedIdea = await ideaDao.updateIdea(req.params.id, req.body);
-    updatedIdea.__v = undefined;
-    updatedIdea.creator = undefined;
-    updatedIdea._id = undefined;
-    res.status(200).send(updatedIdea);
+    const { id } = req.params;
+    const idea = await ideaDao.getIdea(id);
+    console.log(idea);
+    if (idea) {
+      if (idea.creator === req.decoded.userEmail) {
+        req.body.average_score = (req.body.impact + req.body.ease + req.body.confidence) / 3.0;
+        const updatedIdea = await ideaDao.updateIdea(req.params.id, req.body);
+        let finalIdea = {
+          id: updatedIdea._id,
+          ...updatedIdea._doc,
+          _id: undefined,
+          __v : undefined,
+          creator :undefined,
+        }
+        res.status(200).send(finalIdea);
+      }else{
+        res.status(401).send();
+      }
+    }else {
+      res.status(404).send('Idea not found');
+    }
   }
 };
